+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Development
 
+### Changed
+- [Internal] Upgraded `bitmovin player` from `3.78.0` to `3.81.0`
+- [Internal] Upgraded `media3 exoplayer` from `1.4.0` to `1.4.1`
+
 ## v3.8.0
 
 ### Added